♻ **Refactor Load Balancer: Move Load Balancing Logic to ** - proxy.go → Handles request forwarding only - lb.go → Handles backend selection & health checks
